+++ FirebaseCombineSwift/Tests/Unit/FirebaseCombine-unit-Bridging-Header.h
@@ -14,11 +14,23 @@
 
 #import "SharedTestUtilities/FIROptionsMock.h"
 
-#import <FirebaseAuth/FIREmailAuthProvider.h>
+// Firebase Core
+#import "FirebaseCore/Sources/Private/FirebaseCoreInternal.h"
+
+// Firebase Functions
+#import "Functions/FirebaseFunctions/FIRFunctions+Internal.h"
+
+// Firebase Auth
+#import <FirebaseAuth/FIRAuth.h>
+#import <FirebaseAuth/FirebaseAuth.h>
+
 #import "FirebaseAuth/Sources/Auth/FIRAuthGlobalWorkQueue.h"
 #import "FirebaseAuth/Sources/Auth/FIRAuth_Internal.h"
+
+#import "FirebaseAuth/Sources/AuthProvider/FIRAuthCredential_Internal.h"
 #import "FirebaseAuth/Sources/AuthProvider/GameCenter/FIRGameCenterAuthCredential.h"
 #import "FirebaseAuth/Sources/AuthProvider/OAuth/FIROAuthCredential_Internal.h"
+
 #import "FirebaseAuth/Sources/Backend/FIRAuthBackend.h"
 #import "FirebaseAuth/Sources/Backend/RPC/FIRCreateAuthURIRequest.h"
 #import "FirebaseAuth/Sources/Backend/RPC/FIRCreateAuthURIResponse.h"
@@ -57,4 +69,3 @@
 #import "FirebaseAuth/Sources/Utilities/FIRAuthErrorUtils.h"
 #import "FirebaseAuth/Sources/Utilities/FIRAuthURLPresenter.h"
 #import "FirebaseAuth/Sources/Utilities/FIRAuthWebUtils.h"
-#import "FirebaseCore/Sources/Private/FirebaseCoreInternal.h"