+++ FirebaseCombineSwift/CHANGELOG.md
@@ -1,3 +1,4 @@
 # Unreleased
 
+- [feature] Added Combine support for Cloud Functions for Firebase
 - [feature] WIP: Added Combine support for Firebase Auth